+++ src/client/cypress/e2e/helpers/formHelpers.js
@@ -1,11 +1,13 @@
+import { createBaseSelector } from "./testHelpers";
+
 /**
  * Sets the value for an input form element.
  * @param {string} fieldName The name of the input field.
  * @param {string} text The text to type into the input field.
- * @param {boolean} clear Whether to clear the input field before typing
+ * @param {string} parent (optional) The parent of the form element.
  */
-export const setInput = (fieldName, value) => {
-  var selector = `[data-cy="${fieldName}-formInput"]`;
+export const setInput = (fieldName, value, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formInput"]`;
   cy.get(selector)
     .click()
     .then(() => {
@@ -15,16 +17,15 @@ export const setInput = (fieldName, value) => {
       });
     });
 };
-export const evaluateInput = (fieldName, expectedValue) => {
-  var selector = `[data-cy="${fieldName}-formInput"] input`;
-  cy.get(selector)
-    .filter((k, input) => {
-      return input.value === expectedValue;
-    })
-    .should("have.length", 1);
-};
-export const evaluateTextarea = (fieldName, expectedValue) => {
-  var selector = `[data-cy="${fieldName}-formInput"] textarea`;
+
+/**
+ * Evaluates the state of an input form element
+ * @param {string} fieldName The name of the select field.
+ * @param {number} expectedValue The expected number of options in the dropdown.
+ * @param {string} parent (optional) The parent of the form element.
+ */
+export const evaluateInput = (fieldName, expectedValue, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formInput"] input`;
   cy.get(selector)
     .filter((k, input) => {
       return input.value === expectedValue;
@@ -32,20 +33,50 @@ export const evaluateTextarea = (fieldName, expectedValue) => {
     .should("have.length", 1);
 };
 
+/**
+ * Evaluates the state of a textarea form element
+ * @param {string} fieldName The name of the select field.
+ * @param {number} expectedValue The expected number of options in the dropdown.
+ * @param {string} parent (optional) The parent of the form element.
+ */
+export const evaluateTextarea = (fieldName, expectedValue, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formInput"] textarea`;
+  cy.get(selector)
+    .filter((k, input) => {
+      return input.value === expectedValue;
+    })
+    .should("have.length", 1);
+};
+
+/**
+ * Opens the dropdown for a select or multi-select form element.
+ * @param {string} selector The selector for the form element.
+ */
 export const openDropdown = selector => {
   cy.get(selector).find('[role="combobox"]').click();
 };
 
+/**
+ * Selects an option from a dropdown.
+ * @param {number} index The index of the option to select.
+ */
 export const selectDropdownOption = index => {
   cy.get('.MuiPaper-elevation [role="listbox"]').find('[role="option"]').eq(index).click();
 };
 
+/**
+ * Evaluates the number of options in a dropdown.
+ * @param {number} length The expected number of options in the dropdown.
+ */
 export const evaluateDropdownOptionsLength = length => {
   cy.get('.MuiPaper-elevation [role="listbox"]').should($listbox => {
     expect($listbox.find('[role="option"]')).to.have.length(length);
   });
 };
 
+/**
+ * Closes the dropdown for a select or multi-select form element.
+ */
 export const closeDropdown = () => {
   cy.get("body").click();
 };
@@ -55,9 +86,10 @@ export const closeDropdown = () => {
  * @param {string} fieldName The name of the select field.
  * @param {number} index The index of the option to select.
  * @param {number} expected The expected number of options in the dropdown.
+ * @param {string} parent (optional) The parent of the form element.
  */
-export const setSelect = (fieldName, index, expected) => {
-  var selector = `[data-cy="${fieldName}-formSelect"]`;
+export const setSelect = (fieldName, index, expected, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formSelect"]`;
   openDropdown(selector);
   if (expected != null) {
     evaluateDropdownOptionsLength(expected);
@@ -65,8 +97,14 @@ export const setSelect = (fieldName, index, expected) => {
   selectDropdownOption(index);
 };
 
-export const evaluateSelect = (fieldName, expectedValue) => {
-  var selector = `[data-cy="${fieldName}-formSelect"] input`;
+/**
+ * Evaluates the state of a select form element.
+ * @param {string} fieldName The name of the select field.
+ * @param {string} expectedValue The expected value of the select.
+ * @param {string} parent (optional) The parent of the form element.
+ */
+export const evaluateSelect = (fieldName, expectedValue, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formSelect"] input`;
   cy.get(selector)
     .filter((k, input) => {
       return input.value === expectedValue;
@@ -75,13 +113,14 @@ export const evaluateSelect = (fieldName, expectedValue) => {
 };
 
 /**
- * Sets the value for a select form element.
+ * Sets the value for a multi-select form element.
  * @param {string} fieldName The name of the select field.
  * @param {number[]} indices The indices of the options to select.
  * @param {number} expected The expected number of options in the dropdown.
+ * @param {string} parent (optional) The parent of the form element.
  */
-export const toggleMultiSelect = (fieldName, indices, expected) => {
-  var selector = `[data-cy="${fieldName}-formMultiSelect"]`;
+export const toggleMultiSelect = (fieldName, indices, expected, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formMultiSelect"]`;
   openDropdown(selector);
   if (expected != null) {
     evaluateDropdownOptionsLength(expected);
@@ -92,8 +131,14 @@ export const toggleMultiSelect = (fieldName, indices, expected) => {
   closeDropdown();
 };
 
-export const evaluateMultiSelect = (fieldName, expectedValues) => {
-  var selector = `[data-cy="${fieldName}-formMultiSelect"] input`;
+/**
+ * Evaluates the state of a multi-select form element.
+ * @param {string} fieldName The name of the multi-select field.
+ * @param {string[]} expectedValues The expected values of the multi-select.
+ * @param {string} parent (optional) The parent of the form element.
+ */
+export const evaluateMultiSelect = (fieldName, expectedValues, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formMultiSelect"] input`;
   cy.get(selector)
     .filter((k, input) => {
       if (expectedValues.length === 0) {
@@ -109,27 +154,37 @@ export const evaluateMultiSelect = (fieldName, expectedValues) => {
 /**
  * Toggles the checkbox for a checkbox form element.
  * @param {string} fieldName The name of the checkbox field.
+ * @param {string} parent (optional) The parent of the form element.
  */
-export const toggleCheckbox = fieldName => {
-  var selector = `[data-cy="${fieldName}-formCheckbox"]`;
+export const toggleCheckbox = (fieldName, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formCheckbox"]`;
   cy.get(selector).click();
 };
+
 /**
- * Sets the value for a select form element.
+ * Evaluates the state of a checkbox form element.
  * @param {string} fieldName The name of the checkbox field.
  * @param {number} expectedChecked The expected state of the checkbox.
+ * @param {string} parent (optional) The parent of the form element.
  */
-export const evaluateCheckbox = (fieldName, expectedChecked) => {
-  var checked = expectedChecked ? "true" : "false";
-  var selector = `[data-cy="${fieldName}-formCheckbox"]`;
-  cy.get(selector).invoke("attr", "aria-disabled").should("eq", checked);
+export const evaluateCheckbox = (fieldName, expectedChecked, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formCheckbox"]`;
+  cy.get(selector).invoke("attr", "aria-disabled").should("eq", expectedChecked);
 };
-export const evaluateDisplayValue = (fieldName, expectedValue) => {
+
+/**
+ * Evaluates the display value of a form element.
+ * @param {string} fieldName The name of the checkbox field.
+ * @param {number} expectedChecked The expected state of the checkbox.
+ * @param {string} parent (optional) The parent of the form element.
+ */
+export const evaluateDisplayValue = (fieldName, expectedValue, parent) => {
+  var selector = createBaseSelector(parent) + `[data-cy="${fieldName}-formDisplay"]`;
   if (Array.isArray(expectedValue)) {
     expectedValue.forEach(value => {
-      cy.get(`[data-cy="${fieldName}-formDisplay"]`).contains(value);
+      cy.get(selector).contains(value);
     });
   } else {
-    cy.get(`[data-cy="${fieldName}-formDisplay"]`).contains(expectedValue);
+    cy.get(selector).contains(expectedValue);
   }
 };